requirements.txt
Django==4.2
django-crispy-forms
bootstrap4
psycopg2-binary
pillow
crispy_bootstrap5
multiselectfield
django-jet-reboot
google-api-python-client==1.4.1
feedparser